Advanced next card selection to avoid repeats when at all possible.
Note: In this I got rid of ReviewSession and merged it into the document. In hindsight, I'm not sure this was necessary. It looked like multiple instances of ReviewSession were being created, but I may have misunderstood what was happening. Either way, it works now and is arguably cleaner data-flow-wise, so I'm keeping it as is unless I run into issues with saving.