#182891724 - Adjusted animations on Review screen
Last commit I was very focused on the card itself, and then I tested out the review screen and realized it was really buggy when you change cards.

I cleaned this up by adding an animation where the card goes off screen and slides back on when changing cards, like it's going into a deck and a new one is being dealt.

I also adjusted the animations for the buttons to feel as good as I could get them. Sliding in seemed weird for some reason, but sliding out looks great, so I had them fade in and slide out.