#182873114 - Fixed being able to type kanji in edit fields
Issue was because updating text binding lead to the view updating and recreating the text fields. This was resolved by adding a new MirroredTextField View that keeps an internal @State variable to hold the typed value and handles mirroring that to the provided source. It's interface is compatible with the standard override for TextField, so it is a very easy replacement to go back if the bug is fixed in the future.